grpc-js-xds: Factor Fraction out into a separate file
